requirements.txt
requests
python-telegram-bot
flask
google-generativeai
python-dotenv